Release 20160630 - 0.11.7

+++ controllers/imageController.js
@@ -35,24 +35,46 @@ ImageController.prototype.upload = function(req, res, next) {
 		.catch(next)
 }
 
-ImageController.prototype.setUserAvatar = function(req, res, next) {
+ImageController.prototype.getMetadata = function(req, res, next) {
+	var path = req.path.replace(/^\/meta/, '')
+
+	this._service.findOne({'scaled.url' : path}).then(function(item) {
+		if (!item) {
+			return next()
+		}
+
+		res.header('Cache-Control', 'public')
+
+		res.json({
+			width: item.scaled.width,
+			height: item.scaled.height,
+			pitch: item.scaled.pitch,
+			heading: item.scaled.heading,
+			roll: item.scaled.roll
+		})
+	})
+	.catch(next)
+}
+
+// called by setUserAvatar and setUserHeader
+ImageController.prototype._setUserProfileImage = function(req, res, next, imageProcessor, folder, profileFields) {
 	var that = this
 
+	var file = req.files.file
+	if (!(file && file.path))
+		return res.status(400).send('Please upload a file')
+
+
 	User.findById(req.user.id, function(err, user) {
+		
 		if (err || !user)
 			return next(err)
 
-		var file = req.files.file
-		var folder = '/' + req.user.username + '/profile'
-
-		new ImageProcessor(that._fs)
-			.handleAvatarUpload(file, folder)
+		imageProcessor(file, folder)
 			.then(function(info) {
 				fs.unlink(file.path, function() {})
-
-				user.profile.avatarOriginal = info.original.url
-				user.profile.avatarScaled = info.scaled.url
-
+				user.profile[profileFields.original] = info.original.url
+				user.profile[profileFields.scaled] = info.scaled.url
 	 			user.save(function(err) {
 	 				if (err)
 	 					return next(err)
@@ -68,6 +90,26 @@ ImageController.prototype.setUserAvatar = function(req, res, next) {
 			.catch(next)
 	})
 }
+
+
+ImageController.prototype.setUserAvatar = function(req, res, next) {
+	var processor = new ImageProcessor(this._fs)
+	return this._setUserProfileImage(req, res, next,
+		processor.handleUserAvatarUpload.bind(processor),
+		'/' + req.user.username + '/profile/avatar',
+		{original: 'avatarOriginal', scaled: 'avatarScaled'}
+	)
+}
+
+ImageController.prototype.setUserHeader = function(req, res, next) {
+	var processor = new ImageProcessor(this._fs)
+	return this._setUserProfileImage(req, res, next,
+		processor.handleUserHeaderUpload.bind(processor),
+		'/' + req.user.username + '/profile/header',
+		{original: 'headerOriginal', scaled: 'headerScaled'}
+	)
+}
+
 
 ImageController.prototype.uploadAnonymous = function(req, res, next) {
 	var that = this
